Reconstruct binance unit tests
